Added Heart Rate Sensor Location
This was missing! Because of course, everything needs its own variation of sensor location!
